+++ src/app/app.module.ts
@@ -9,7 +9,7 @@ import { MatSlideToggleModule } from "@angular/material/slide-toggle";
 
 import { AppRoutingModule } from "./app-routing.module";
 import { AppComponent } from "./app.component";
-import { SignUpComponent } from "./sign-up/sign-up.component";
+import { SignUpComponent } from "./auth/pages/signup/signup.component";
 import { ProfileComponent } from "./profile/profile.component";
 
 type ImportedModule = Type<any> | ModuleWithProviders<{}> | any[];
