+++ src/app/auth/pages/signup/signup.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { SignUpComponent } from "./sign-up.component";
+import { SignUpComponent } from "./signup.component";
 
 describe("SignUpComponent", () => {
   let component: SignUpComponent;
