+++ src/app/auth/pages/signup/signup.component.ts
@@ -1,0 +1,10 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "app-signup",
+  templateUrl: "./signup.component.html",
+  styleUrls: ["./signup.component.css"],
+})
+export class SignUpComponent {
+  constructor() {}
+}
