--- src/app/sign-up/sign-up.component.ts
@@ -1,10 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-  selector: "app-sign-up",
-  templateUrl: "./sign-up.component.html",
-  styleUrls: ["./sign-up.component.css"],
-})
-export class SignUpComponent {
-  constructor() {}
-}
